+++ app_plus/src/main/java/org/gkisalatiga/plus/fragment/FragmentSeasonalAgenda.kt
@@ -11,15 +11,42 @@
 package org.gkisalatiga.plus.fragment
 
 import androidx.activity.ComponentActivity
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import org.gkisalatiga.plus.data.ActivityData
+import org.gkisalatiga.plus.db.ModulesCompanion
 
 class FragmentSeasonalAgenda (private val current : ActivityData) : ComponentActivity() {
 
+    // The root seasonal node.
+    private val seasonalNode = ModulesCompanion.jsonRoot!!.getJSONObject("seasonal")
+
     @Composable
     fun getComposable() {
-        Text("Testing Agenda")
+        Column {
+            // Display the main title.
+            val mainTitle = seasonalNode.getJSONObject("static-menu").getJSONObject("agenda").getString("title")
+            Row (verticalAlignment = Alignment.CenterVertically, horizontalArrangement = Arrangement.Center, modifier = Modifier.fillMaxWidth()) {
+                Text(mainTitle, modifier = Modifier, fontWeight = FontWeight.Bold, fontSize = 24.sp, overflow = TextOverflow.Ellipsis, textAlign = TextAlign.Center)
+            }
+
+            Spacer(Modifier.height(20.dp))
+
+            Text("Testing Agenda")
+        }
     }
 
 }